macos: M2 (part 1) — re-base 32-bit address fields on the boot path
First batch of the M2 sweep (docs/dev/MACOS-ARM64-PLAN.md): every site below
stores an N64 DRAM address in a 32-bit field and later casts it back to a
pointer. On Windows/Linux the N64 address *is* the host address, so the raw
cast works; on native arm64 the address space is shifted by PORT_ADDR_BASE, so
the cast must go through portN64ToHost. All edits are identity at
PORT_ADDR_BASE == 0 and guarded #if defined(PORT), so the N64 and
Windows/Linux builds are unchanged.

Boot progress: the binary now boots past ROM load, language banks, the
animation tables and into the level stage load (texReset/texLoad), where the
next un-rebased site faults. Before this batch it faulted in langInit.

- memp.c: MemoryPool fields are u8*, so convert once at the pool boundary
  (mempCheckMemflagTokens / mempSetBankStarts) instead of at every allocation.
  The s32 addresses the game passes around are kept s32 (V1 must stay
  s32-positive for mempCheckMemflagTokens).
- language.c: g_LangBanks[] slots (s32) re-based at their 9 pointer uses, and
  the D129 plausibility guard's bounds move with the base.
- initanitable.c: expand_ani_table_entries() builds N64 addresses in s32 vars;
  re-base the three derefs.
- initactorpropstuff.c: weapon_firing_animation_table.anim, StruckAnim, and
  the ANIM_PTR/ANIM_FRAC macros re-based (the union stores a pointer but is
  filled with an N64 address).
- model.c: ModelAnimation.bitDescriptors/bitStream (D32 u32 address fields)
  re-based at modelAnimReadRootMotionValue's two uses.
- image_bank.c/h: globalbank_rdram_offset becomes host-based under PORT, so
  `globalbank_rdram_offset + GIMG_OFF(sym)` is a live pointer for all ~43
  texReset uses at once (`- 0x02000000` is the 64-bit equivalent of the
  original mod-2^32 `+ 0xFE000000` cancellation); pGlobalimagetable re-based
  for romCopy/gimgFixup.
- image.c: texLoadFromDisplayList's `(u32*)((s32)bytes + 4)` truncation →
  `bytes + 4` (identical on 32-bit/64-bit base-0).

docs/porting-notes.md: new §F documents the whole class (symptom, fix, grep
heuristic) for the next agent, per the project's porting-notes discipline.

+++ src/game/image.c
@@ -2462,7 +2462,7 @@ void texLoadFromDisplayList(Gfx *gdl, struct texpool *arg1)
         if (bytes[0] == G_SETTIMG && bytes[4] == 0xab && bytes[5] == 0xcd)
 #endif
         {
-            texLoad((u32 *)((s32)bytes + 4), arg1);
+            texLoad((u32 *)(bytes + 4), arg1);
         }
 
         bytes += 8;
